+++ run_taskbar.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python3 
+# -*- coding: utf-8 -*-
+
+
 #    Copyright (C) 2019  Marcus Rickert
 #
 #    See https://github.com/marcus67/little_brother
@@ -16,28 +20,7 @@
 #    with this program; if not, write to the Free Software Foundation, Inc.,
 #    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-import os
+from little_brother import taskbar_app
 
-from little_brother import audio_handler
-from little_brother import mpg123_audio_player
-#from little_brother import playsound_audio_player
-from python_base_app.test import base_test
-
-HELLO_MPG = os.path.join(os.path.dirname(__file__), "resources/hello.mpg")
-
-
-class TestAudioPlayer(base_test.BaseTestCase):
-
-    @base_test.skip_if_env("NO_AUDIO_OUTPUT")
-    def test_mpg123_player(self):
-
-        a_config = audio_handler.AudioHandlerConfigModel()
-        a_player = mpg123_audio_player.Mpg123AudioPlayer(p_mpg123_binary = a_config.mpg123_binary)
-        a_player.play_audio_file(HELLO_MPG)
-
-    # @base_test.skip_if_env("NO_AUDIO_OUTPUT")
-    # def test_playsound_player(self):
-    #
-    #     a_player = playsound_audio_player.PlaysoundAudioPlayer()
-    #     a_player.play_audio_file(HELLO_MPG)
-
+if __name__ == '__main__':
+    exit(taskbar_app.main())